requirements.txt
aiohttp
python-dotenv
yt-dlp
